Handle search requests uniformly to use filters and ordering
